requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
graphviz==0.14.1
identify==1.5.13
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.5.0
Pillow==8.0.1
pre-commit==2.9.3
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
svgwrite==1.3.1
toml==0.10.2
urllib3==1.26.2
virtualenv==20.4.0
Werkzeug==1.0.1
